# apiDocumentation.yaml
swagger: '2.0'
info:
  description: ADES CA1
  version: 1.0.0
  title: ADES
  # put the contact info for your development or API team
  contact:
    email: zhengwei.20@ichat.sp.edu.sg

  license:
    name: Website URL
    url: https://lzwades.herokuapp.com
paths:
...
  /user/login:
    post:
      summary: user login
      operationId: login
      description: |
        user login
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: detail
        description: user detail
        required: true
        schema: 
          $ref: '#/definitions/UserLogin'
      responses:
        200:
          description: return order history
          schema:
            $ref: '#/definitions/UserLoginResult'
        401: 
          description: Verify user error
        500:
          description: Internal Server Error
  /user/register:
    post:
      summary: user register
      operationId: register
      description: |
        user login
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: detail
        description: user detail
        required: true
        schema: 
          $ref: '#/definitions/UserRegister'
      responses:
        200:
          description: return order history
          schema:
            $ref: '#/definitions/UserRegisterResult'
        401: 
          description: Verify user error
        500:
          description: Internal Server Error
  /user/checkLogin:
    post:
      summary: check user login
      operationId: check login
      description: |
        check user login
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: detail
        description: user detail
        required: true
        schema: 
          $ref: '#/definitions/UserVerify'
      responses:
        200:
          description: return user name
          schema:
            $ref: '#/definitions/UserDetail'
        401: 
          description: Verify user error
        500:
          description: Internal Server Error
...
  /user/forgetPass:
    post:
      summary: user forget password
      operationId: forgetPass
      description: |
        check user login
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        required: true
        name: email
        description: user email
        schema: 
          $ref: '#/definitions/UserEmail'
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserForgetPassResult'
        400: 
          description: Wrong email format
        404: 
          description: User email does not exist
        500:
          description: Internal Server Error
  /user/verifyResetPass:
    post:
      summary: user reset password
      operationId: verifyResetPass
      description: |
        user reset password
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        required: true
        name: detail
        description: user detail
        schema: 
          $ref: '#/definitions/UserResetPassDetail'
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserForgetPassResult'
        400: 
          description: Wrong password format, wrong code
        500:
          description: Internal Server Error
  /user/verifyEmail:
    post:
      summary: user register verify email
      operationId: verifyEmail
      description: |
        user register verify email
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: email
        required: true
        description: user email
        schema: 
          $ref: '#/definitions/UserEmail'
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserForgetPassResult'
        400: 
          description: email already registered
        500:
          description: Internal Server Error
  /user/saveUserInfo:
    post:
      summary: user update detail
      operationId: saveUserInfo
      description: |
        user update detail
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        required: true
        name: user detail
        description: user detail
        schema: 
          $ref: '#/definitions/UserUpdateDetail'
      - in: header
        name: token
        description: user login token
        type: string
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserRegisterResult'
        204: 
          description: no user to update
        500:
          description: Internal Server Error
  /user/supportRequest:
    post:
      summary: user request support
      operationId: supportRequest
      description: |
        user request support
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: detail
        required: true
        description: detail
        schema: 
          $ref: '#/definitions/UserRequestSupport'
      - in: header
        name: token
        description: user login token
        required: true
        type: string
      responses:
        200:
          description: return status
          schema:
            $ref: '#/definitions/UserRequestSupportResult'
        500:
          description: Internal Server Error
...
definitions:
  UserVerify:
    type: object
    required:
    - uid
    properties:
      uid:
        type: integer
        format: int64
        example: 1820797504
...
  UserLogin:
    type: object
    required:
    - email
    - password
    - captcha
    - secretCode
    properties:
      email:
        type: string
        example: example@gamil.com
      password:
        type: string
        example: example
      captcha: 
        type: string
        example: captcha
      secretCode:
        type: string
        example: 012345
  UserLoginResult:
    type: object
    required:
    - token
    - success
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywibmFtZSI6IkxpbHkiLCJpYXQiOjE2Mzc5MDI3MjIsImV4cCI6MTYzNzk4OTEyMn0.sybmhllBjRafuPKMlyxnTmuCs2JToI0oF85sC9gmXIg
      success:
        type: string
        example: true
  UserRegister:
    type: object
    required:
    - username
    - useremail
    - userpassword
    - userphone
    - usergender
    - code
    properties:
      username:
        type: string
        example: user name
      useremail:
        type: string
        example: lily@gmail.com
      userpassword:
        type: string
        example: passwordInput
      userphone: 
        type: string
        example: 01234567
      usergender:
        type: string
        example: 1
      code:
        type: string
        example: code
  UserRegisterResult:
    type: object
    required:
    - token
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywibmFtZSI6IkxpbHkiLCJpYXQiOjE2Mzc5MDI3MjIsImV4cCI6MTYzNzk4OTEyMn0.sybmhllBjRafuPKMlyxnTmuCs2JToI0oF85sC9gmXIg
  UserDetail:
    type: object
    required:
    - name
    - id
    - email
    - phone
    - gender
    - role
    properties:
      name:
        type: string
        example: lily
      id:
        type: string
        example: 1
      email:
        type: string
        example: lily@gmail.com
      phone:
        type: string
        example: 01234567
      gender:
        type: string
        example: 1
      role:
        type: string
        example: 0
  UserEmail:
    type: object
    required:
    - email
    properties:
      email:
        type: string
        example: lily@gmail.com
  UserForgetPassResult:
    type: object
    required:
    - status
    properties:
      status:
        type: string
        example: done
  UserResetPassDetail:
    type: object
    required:
    - email
    - password
    - code
    properties:
      email: 
        type: string
        example: lily@gmail.com
      password:
        type: string
        example: password
      code:
        type: string
        example: qwertyuiop0987654321
  UserUpdateDetail:
    type: object
    required:
    - username
    - phone
    - gender
    properties:
      username:
        type: string
        example: Lily
      phone:
        type: string
        example: 01234567
      gender:
        type: string
        example: 1
  UserRequestSupport:
    type: object
    required:
    - email
    - subject
    - message
    properties:
      email:
        type: string
        example: lily@gmail.com
      subject:
        type: string
        example: subject
      message:
        type: string
        example: message
  UserRequestSupportResult:
    type: object
    required:
    - status
    properties:
      status:
        type: string
        example: done
...
host: virtserver.swaggerhub.com
basePath: /lzwades/ades/1.0.0
schemes:
 - https